src: discord: newShard: prepend token with 'Bot '
Don't expect users to call 'newShard' like 'newShard("Bot <token>")'
but instead offer an simpler version: 'newShard("<token>")' and take
care of the string on your own.

The only case where this can fail is if there can be something else than
"Bot ". For this case there is the "kind" parameter. Where this can be
changed.